Iteration 2 - improve byline
In this iteration, we extend the functionality of our Python module by introducing a new function, `get_byline()`, which returns the value of a global variable, `byline`. This update enhances the modularity and testability of our code. By isolating the retrieval of the byline into its own function, we make the code more flexible and easier to maintain.